+++ src/queries/useAuthoritiesBrowse/useAuthoritiesBrowse.js
@@ -68,6 +68,7 @@ const useBrowseRequest = ({
     }, {
       keepPreviousData: true,
       staleTime: 5 * 60 * 1000,
+      cacheTime: 0,
     },
   );
 
@@ -148,7 +149,6 @@ const useAuthoritiesBrowse = ({
 }) => {
   const [currentQuery, setCurrentQuery] = useState(searchQuery);
   const [currentIndex, setCurrentIndex] = useState(searchIndex);
-  const [hasEmptyAnchor, setHasEmptyAnchor] = useState(false);
   const [items, setItems] = useState([]);
   const {
     page,
@@ -203,12 +203,8 @@ const useAuthoritiesBrowse = ({
     setItems(mainRequest.data?.items || []);
   }, [mainRequest.data]);
 
-  useEffect(() => {
-    if (page === 0) {
-      const dataIncludesEmptyAnchor = !!mainRequest.data?.items.find(item => !item.authority);
-
-      setHasEmptyAnchor(dataIncludesEmptyAnchor);
-    }
+  const hasEmptyAnchor = useMemo(() => {
+    return page === 0 && totalRecords !== 0 && !!mainRequest.data?.items.find(item => !item.authority);
   }, [mainRequest.data]);
 
   const itemsWithPrevAndNextPages = useMemo(() => {